feat: navBar and components
