+++ src/app/components/Footer.tsx
@@ -1,0 +1,11 @@
+const Footer = () => (
+  <footer className="max-w-4xl mx-auto p-4">
+    <div className="container mx-auto text-center">
+      <p className="text-sm">
+        &copy; {new Date().getFullYear()} Pulppo. Todos los derechos reservados.
+      </p>
+    </div>
+  </footer>
+);
+
+export default Footer;
