+++ src/app/layout.tsx
@@ -1,12 +1,15 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
+
+import Header from "./components/Header";
+
 import "./globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
   title: "Pulppo Challenge",
-  description: "Challenge para nuevos developers del equipo de pulppo",
+  description: "Challenge realizado por Pablo Achaval",
 };
 
 export default function RootLayout({
@@ -16,7 +19,10 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <Header />
+        <main className="max-w-4xl mx-auto p-4">{children}</main>
+      </body>
     </html>
   );
 }
